etu_fastapi_frame/fastapi/0.95.2/ms/requirements.txt
fastapi[all]==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.8
PyJWT==2.7.0
cryptography==41.0.1
passlib==1.7.4
bcrypt==4.0.1
